pyproject.toml
[tool.poetry]
name = "task-tracker-api"
version = "0.1.0"
description = ""
authors = [
    "BuilderCrafter <Aleksa.gb317@gmail.com>"
]
readme = "README.md"
packages = [
    { include = "app"} 
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
passlib = { version = ">=1.7.4,<2.0.0", extras = ["bcrypt"] }
uvicorn = { version = ">=0.34.3,<0.35.0", extras = ["standard"] }
sqlalchemy = ">=2.0.41,<3.0.0"
asyncpg = ">=0.30.0,<0.31.0"
pydantic-settings = ">=2.9.1,<3.0.0"
fastapi = ">=0.115.13,<0.116.0"
pydantic = { version = ">=2.11.7,<3.0.0", extras = ["email"] }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"
ruff = "^0.11.13"
black = "^25.1.0"
mypy = "^1.16.0"
coverage = "^7.9.1"
pre-commit = "^4.2.0"
pydantic = {extras = ["mypy"], version = "^2.11.7"}
pydantic-settings = "^2.10.0"
alembic = "^1.16.2"
types-passlib = "^1.7.7.20250602"
factory-boy = "^3.3.3"
aiosqlite = "^0.21.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
testcontainers = {extras = ["postgres"], version = "^4.10.0"}
factory-boy = "^3.3.3"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"

[tool.mypy]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

